got rid of some warnings and made the map bigger

+++ src/camera.h
@@ -49,15 +49,15 @@ void mouse_callback(GLFWwindow* window, double xpos, double ypos)
 {
     if (firstMouse)
     {
-        lastX = xpos;
-        lastY = ypos;
+        lastX = (float) xpos;
+        lastY = (float) ypos;
         firstMouse = false;
     }
 
-    float xoffset = xpos - lastX;
-    float yoffset = lastY - ypos;
-    lastX = xpos;
-    lastY = ypos;
+    float xoffset = (float) xpos - lastX;
+    float yoffset = lastY - (float) ypos;
+    lastX = (float) xpos;
+    lastY = (float) ypos;
 
     float sensitivity = 0.1f;
     xoffset *= sensitivity;
